Extract common stuff to common.py and merge data_copier.py with data.py
