+++ neuro_extras/common.py
@@ -1,0 +1,2 @@
+NEURO_EXTRAS_IMAGE_TAG = "v20.9.30a7"
+NEURO_EXTRAS_IMAGE = f"neuromation/neuro-extras:{NEURO_EXTRAS_IMAGE_TAG}"
